LayoutTests/fast/backgrounds/background-shorthand-with-backgroundSize-style.html

<!DOCTYPE html>
<html>
<script src="../../resources/js-test-pre.js"></script>
<div id="test"> </div>
<script>
description("Tests background shortand property with background-size");

var e = document.getElementById('test');
var computedStyle = getComputedStyle(e);

function checkStyle() {
    var before = e.style.background;
    e.style.background = 'none';
    e.style.background = before;
    styleAfterRoundTrip = e.style.background;
    shouldBeEqualToString("styleAfterRoundTrip", before);
}

function checkComputedStyleValue() {
    var before = computedStyle.background;
    e.style.background = 'none';
    e.style.background = before;
    computedStyleAfterRoundTrip = computedStyle.background;
    shouldBeEqualToString("computedStyleAfterRoundTrip", before);
}

e.style.background = "center / cover red url(dummy://test.png) no-repeat border-box";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") center center / cover no-repeat border-box');
shouldBeEqualToString("e.style.backgroundSize", 'cover');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 50% 50% / cover no-repeat border-box');
shouldBeEqualToString('computedStyle.backgroundSize', 'cover');
checkComputedStyleValue()
debug("")

e.style.background = "red 20px / contain url(dummy://test.png) no-repeat padding-box";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") 20px center / contain no-repeat padding-box');
shouldBeEqualToString("e.style.backgroundSize", 'contain');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 20px 50% / contain no-repeat padding-box');
shouldBeEqualToString('computedStyle.backgroundSize', 'contain');
checkComputedStyleValue()
debug("")

e.style.background = "red url(dummy://test.png) 50px 60px / 50% 75% no-repeat";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") 50px 60px / 50% 75% no-repeat');
shouldBeEqualToString("e.style.backgroundSize", '50% 75%');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 50px 60px / 50% 75% no-repeat');
shouldBeEqualToString('computedStyle.backgroundSize', '50% 75%');
checkComputedStyleValue()
debug("")

e.style.background = "red url(dummy://test.png) repeat top left / 100px 200px border-box content-box";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") left top / 100px 200px border-box content-box');
shouldBeEqualToString("e.style.backgroundSize", '100px 200px');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 0% 0% / 100px 200px border-box content-box');
shouldBeEqualToString('computedStyle.backgroundSize', '100px 200px');
checkComputedStyleValue()
debug("")

e.style.background = "red url(dummy://test.png) repeat 50% / auto auto content-box padding-box";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") 50% content-box padding-box');
shouldBeEqualToString("e.style.backgroundSize", 'auto');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 50% 50% content-box padding-box');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")

e.style.background = "url(dummy://test.png) red 50px 60px / 50% no-repeat fixed";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") 50px 60px / 50% no-repeat fixed');
shouldBeEqualToString("e.style.backgroundSize", '50%');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 50px 60px / 50% no-repeat fixed');
shouldBeEqualToString('computedStyle.backgroundSize', '50%');
checkComputedStyleValue()
debug("")

e.style.background = "red repeat scroll padding-box border-box top left / 100px url(dummy://test.png)";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") left top / 100px');
shouldBeEqualToString("e.style.backgroundSize", '100px');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 0% 0% / 100px');
shouldBeEqualToString('computedStyle.backgroundSize', '100px');
checkComputedStyleValue()
debug("")

e.style.background = "50% / auto fixed url(dummy://test.png) repeat content-box red";
shouldBeEqualToString("e.style.background", 'red url("dummy://test.png") 50% fixed content-box');
shouldBeEqualToString("e.style.backgroundSize", 'auto');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) url("dummy://test.png") 50% 50% fixed content-box');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")

e.style.background = "top left / 50%";
shouldBeEqualToString("e.style.background", 'left top / 50%');
shouldBeEqualToString("e.style.backgroundSize", '50%');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgba(0, 0, 0, 0) 0% 0% / 50%');
shouldBeEqualToString('computedStyle.backgroundSize', '50%');
checkComputedStyleValue()
debug("")

e.style.background = "red fixed";
shouldBeEqualToString("e.style.background", 'red fixed');
shouldBeEqualToString("e.style.backgroundSize", 'auto');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgb(255, 0, 0) fixed');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")

e.style.background = "";
e.style.background = "red / cover url(dummy://test.png) repeat";
shouldBeEqualToString("e.style.background", '');
shouldBeEqualToString("e.style.backgroundSize", '');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgba(0, 0, 0, 0)');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")

e.style.background = "";
e.style.background = "red url(dummy://test.png) repeat 25px / contain contain";
shouldBeEqualToString("e.style.background", '');
shouldBeEqualToString("e.style.backgroundSize", '');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgba(0, 0, 0, 0)');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")

e.style.background = "";
e.style.background = "red top left / 100px cover url(dummy://test.png) repeat";
shouldBeEqualToString("e.style.background", '');
shouldBeEqualToString("e.style.backgroundSize", '');
checkStyle()
shouldBeEqualToString('computedStyle.background', 'rgba(0, 0, 0, 0)');
shouldBeEqualToString('computedStyle.backgroundSize', 'auto');
checkComputedStyleValue()
debug("")
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
