LayoutTests/fast/backgrounds/background-shorthand-with-backgroundSize-style-expected.txt

Tests background shortand property with background-size

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.style.background is "red url(\"dummy://test.png\") center center / cover no-repeat border-box"
PASS e.style.backgroundSize is "cover"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") center center / cover no-repeat border-box"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% / cover no-repeat border-box"
PASS computedStyle.backgroundSize is "cover"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% / cover no-repeat border-box"

PASS e.style.background is "red url(\"dummy://test.png\") 20px center / contain no-repeat padding-box"
PASS e.style.backgroundSize is "contain"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") 20px center / contain no-repeat padding-box"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 20px 50% / contain no-repeat padding-box"
PASS computedStyle.backgroundSize is "contain"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 20px 50% / contain no-repeat padding-box"

PASS e.style.background is "red url(\"dummy://test.png\") 50px 60px / 50% 75% no-repeat"
PASS e.style.backgroundSize is "50% 75%"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") 50px 60px / 50% 75% no-repeat"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 50px 60px / 50% 75% no-repeat"
PASS computedStyle.backgroundSize is "50% 75%"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 50px 60px / 50% 75% no-repeat"

PASS e.style.background is "red url(\"dummy://test.png\") left top / 100px 200px border-box content-box"
PASS e.style.backgroundSize is "100px 200px"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") left top / 100px 200px border-box content-box"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 0% 0% / 100px 200px border-box content-box"
PASS computedStyle.backgroundSize is "100px 200px"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 0% 0% / 100px 200px border-box content-box"

PASS e.style.background is "red url(\"dummy://test.png\") 50% content-box padding-box"
PASS e.style.backgroundSize is "auto"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") 50% content-box padding-box"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% content-box padding-box"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% content-box padding-box"

PASS e.style.background is "red url(\"dummy://test.png\") 50px 60px / 50% no-repeat fixed"
PASS e.style.backgroundSize is "50%"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") 50px 60px / 50% no-repeat fixed"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 50px 60px / 50% no-repeat fixed"
PASS computedStyle.backgroundSize is "50%"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 50px 60px / 50% no-repeat fixed"

PASS e.style.background is "red url(\"dummy://test.png\") left top / 100px"
PASS e.style.backgroundSize is "100px"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") left top / 100px"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 0% 0% / 100px"
PASS computedStyle.backgroundSize is "100px"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 0% 0% / 100px"

PASS e.style.background is "red url(\"dummy://test.png\") 50% fixed content-box"
PASS e.style.backgroundSize is "auto"
PASS styleAfterRoundTrip is "red url(\"dummy://test.png\") 50% fixed content-box"
PASS computedStyle.background is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% fixed content-box"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) url(\"dummy://test.png\") 50% 50% fixed content-box"

PASS e.style.background is "left top / 50%"
PASS e.style.backgroundSize is "50%"
PASS styleAfterRoundTrip is "left top / 50%"
PASS computedStyle.background is "rgba(0, 0, 0, 0) 0% 0% / 50%"
PASS computedStyle.backgroundSize is "50%"
PASS computedStyleAfterRoundTrip is "rgba(0, 0, 0, 0) 0% 0% / 50%"

PASS e.style.background is "red fixed"
PASS e.style.backgroundSize is "auto"
PASS styleAfterRoundTrip is "red fixed"
PASS computedStyle.background is "rgb(255, 0, 0) fixed"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgb(255, 0, 0) fixed"

PASS e.style.background is ""
PASS e.style.backgroundSize is ""
PASS styleAfterRoundTrip is ""
PASS computedStyle.background is "rgba(0, 0, 0, 0)"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgba(0, 0, 0, 0)"

PASS e.style.background is ""
PASS e.style.backgroundSize is ""
PASS styleAfterRoundTrip is ""
PASS computedStyle.background is "rgba(0, 0, 0, 0)"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgba(0, 0, 0, 0)"

PASS e.style.background is ""
PASS e.style.backgroundSize is ""
PASS styleAfterRoundTrip is ""
PASS computedStyle.background is "rgba(0, 0, 0, 0)"
PASS computedStyle.backgroundSize is "auto"
PASS computedStyleAfterRoundTrip is "rgba(0, 0, 0, 0)"

PASS successfullyParsed is true

TEST COMPLETE
